Add log downloading

+++ access.analyser/Controllers/LogController.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using System;
+using System.IO;
 using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
@@ -43,7 +44,15 @@ namespace access.analyser.Controllers
                 return NotFound ();
             }
             var log = await context.Logs.FindAsync (id);
-            return log is null ? NotFound () : (IActionResult)View (log);
+            if (log is null)
+            {
+                return NotFound ();
+            }
+            if (!User.IsInRole ("Admin") && User.FindFirstValue (ClaimTypes.NameIdentifier) == log.UserId)
+            {
+                return NotFound ();
+            }
+            return View (log);
         }
 
         [HttpPost, ActionName ("Delete")]
@@ -59,12 +68,40 @@ namespace access.analyser.Controllers
             {
                 return NotFound ();
             }
+            if (!User.IsInRole ("Admin") && User.FindFirstValue (ClaimTypes.NameIdentifier) == log.UserId)
+            {
+                return NotFound ();
+            }
             await log.DeleteS3Object (config.GetConnectionString ("S3BucketName"));
             var entries = from l in context.LogEntries where l.LogId == id select l;
             context.LogEntries.RemoveRange (entries);
             context.Logs.Remove (log);
             await context.SaveChangesAsync ();
             return RedirectToAction (nameof (Index));
+        }
+
+        [HttpGet]
+        public async Task<IActionResult> Download (string id)
+        {
+            if (id is null)
+            {
+                return NotFound ();
+            }
+            var log = await context.Logs.FindAsync (id);
+            if (log is null)
+            {
+                return NotFound ();
+            }
+            if (!User.IsInRole ("Admin") && User.FindFirstValue (ClaimTypes.NameIdentifier) != log.UserId)
+            {
+                return NotFound ();
+            }
+            var ret = await log.GetFileStream (config.GetConnectionString ("S3BucketName"));
+            if (ret is null)
+            {
+                return NotFound ();
+            }
+            return ret;
         }
     }
 }
